tried a poisson solve
